backend/pyproject.toml
[project]
name = "carousel-studio-backend"
version = "0.1.0"
description = "Carousel Studio — local FastAPI backend that renders TikTok carousels from news for 5 topics × multiple designs."
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "feedparser>=6.0",
    "pyyaml>=6.0",
    "pytz>=2024.1",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
]

[tool.setuptools]
packages = ["core", "core.parsers", "designs", "api"]
